:sparkles: Changing to prefLabel.edge_ngram and adding new mode=search. Re-instating code for suggest on multi type. Removing allowMultiple from swagger yml, and adding array type for param + string items
